Remove testcase fee.
